fix: replace persistent term usage with ets table
Types are stored in a public ets table instead of in the persistent
terms. It feels safer since types could be updated and pterm can
get slower with more elements in it so forcing the user to already
have a bunch of terms in it just for this seems like a misuse.